requirements.txt
selenium>=4.10
webdriver-manager>=4.0.1
python-dotenv>=1.0.1
